Migrate cross-trigger links from cue numbers to stable IDs
LightingData.triggers_audio_cue and AudioData.triggers_lighting_cue
now store Option<u32> stable cue IDs instead of Option<f32> numbers.
All resolution paths (fire_audio/lighting_cross_trigger, frame update
loop, circular-check helpers, pending_lighting_triggers queue) updated
to use find_by_id(). Properties panels replaced text-input fields with
ComboBox dropdowns showing "Q{num} {label}" for each eligible cue.
Cue list info column now resolves the ID back to a number for display.

+++ src/audio/playback.rs
@@ -51,7 +51,7 @@ struct ActiveAudioStream {
 pub struct AudioPlaybackEngine {
     streams: Vec<ActiveAudioStream>,
     /// Cross-triggers queued at start() time, drained each frame by app.rs.
-    pending_lighting_triggers: Vec<f32>,
+    pending_lighting_triggers: Vec<u32>,
 }
 
 impl AudioPlaybackEngine {
@@ -276,7 +276,7 @@ impl AudioPlaybackEngine {
     }
 
     /// Drain audio→lighting cross-triggers queued since last call.
-    pub fn take_pending_lighting_triggers(&mut self) -> Vec<f32> {
+    pub fn take_pending_lighting_triggers(&mut self) -> Vec<u32> {
         std::mem::take(&mut self.pending_lighting_triggers)
     }
 }
