+++ src/cue/types.rs
@@ -16,8 +16,8 @@ pub struct LightingData {
     /// Channel intensity values (0-100). Only non-zero channels stored.
     #[serde(deserialize_with = "crate::serde_helpers::deserialize_channel_map")]
     pub channel_values: HashMap<u16, u8>,
-    #[serde(default, serialize_with = "crate::serde_helpers::round_option_f32_2")]
-    pub triggers_audio_cue: Option<f32>,
+    #[serde(default)]
+    pub triggers_audio_cue: Option<u32>,
 }
 
 impl Default for LightingData {
@@ -59,8 +59,8 @@ pub struct AudioData {
     pub fade_out: f32,
     #[serde(default)]
     pub notes: String,
-    #[serde(default, serialize_with = "crate::serde_helpers::round_option_f32_2")]
-    pub triggers_lighting_cue: Option<f32>,
+    #[serde(default)]
+    pub triggers_lighting_cue: Option<u32>,
     /// Fixed playback duration in seconds. When elapsed the cue fades out (or stops if fade_out==0).
     /// None = play until the file ends.
     #[serde(default, serialize_with = "crate::serde_helpers::round_option_f32_2")]
